+++ src/utils/helpers/browser/browser.js
@@ -19,10 +19,10 @@ const Browser = {
   /**
   * Redirect to URL
   *
-  * @method redirectUrl
+  * @method redirectTo
   * @param url => URL string format
   */
-  redirectUrl: (url) => {
+  redirectTo: (url) => {
     Browser.getWindow().location = url;
   },