+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 0.18.0
 
+## !! BREAKING CHANGE !!
+
+* Renamed Browser `redirectUrl` method to `redirectTo`
+
 ## Link (React Router)
 
 Our Link component now supports the React Router. Instead of passing a `href` prop, pass a `to` prop and it will use React Router to navigate.
@@ -9,6 +13,7 @@ Our Link component now supports the React Router. Instead of passing a `href` pr
 * Pod can now receive a prop of `onEdit` - if this is a String it will use it as a `to` prop on a Link component, if it is a Function it will apply it as an `onClick` handler, if it is an object it will apply it's props to the Link.
 * Pod has an additional padding size added of `extra-large`.
 * Pod now applies any additional props to it's top most child element.
+>>>>>>> master
 
 ## Content Updates
 
@@ -49,7 +54,7 @@ Added a redirect action made by the browser. It is now easier to redirect to url
 ```
 import Browser from 'carbon/lib/utils/helpers/browser';
 
-Browser.redirectTo(url)
+Browser.redirectUrl(url)
 ```
 
 # 0.15.0