+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Our Link component now supports the React Router. Instead of passing a `href` pr
 * Pod can now receive a prop of `onEdit` - if this is a String it will use it as a `to` prop on a Link component, if it is a Function it will apply it as an `onClick` handler, if it is an object it will apply it's props to the Link.
 * Pod has an additional padding size added of `extra-large`.
 * Pod now applies any additional props to it's top most child element.
+* We have added a tertiary pod theme.
 
 # 0.17.1